+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/webview/in_app_webview/InAppWebViewClient.java
@@ -120,7 +120,7 @@ public class InAppWebViewClient extends WebViewClient {
     if (isForMainFrame) {
       // There isn't any way to load an URL for a frame that is not the main frame,
       // so call this only on main frame.
-      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
+      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && headers != null)
         webView.loadUrl(url, headers);
       else
         webView.loadUrl(url);
@@ -178,11 +178,7 @@ public class InAppWebViewClient extends WebViewClient {
 
     if (!WebViewFeature.isFeatureSupported(WebViewFeature.DOCUMENT_START_SCRIPT)) {
       String source = webView.userContentController.generateWrappedCodeForDocumentStart();
-      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-        webView.evaluateJavascript(source, (ValueCallback<String>) null);
-      } else {
-        webView.loadUrl("javascript:" + source.replaceAll("[\r\n]+", ""));
-      }
+      webView.evaluateJavascript(source, (ValueCallback<String>) null);
     }
   }
 
@@ -191,11 +187,7 @@ public class InAppWebViewClient extends WebViewClient {
 
     if (!WebViewFeature.isFeatureSupported(WebViewFeature.DOCUMENT_START_SCRIPT)) {
       String source = webView.userContentController.generateWrappedCodeForDocumentEnd();
-      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-        webView.evaluateJavascript(source, (ValueCallback<String>) null);
-      } else {
-        webView.loadUrl("javascript:" + source.replaceAll("[\r\n]+", ""));
-      }
+      webView.evaluateJavascript(source, (ValueCallback<String>) null);
     }
   }
 
@@ -239,12 +231,7 @@ public class InAppWebViewClient extends WebViewClient {
     }
 
     String js = JavaScriptBridgeJS.PLATFORM_READY_JS_SOURCE();
-
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-      webView.evaluateJavascript(js, (ValueCallback<String>) null);
-    } else {
-      webView.loadUrl("javascript:" + js.replaceAll("[\r\n]+", ""));
-    }
+    webView.evaluateJavascript(js, (ValueCallback<String>) null);
 
     if (webView.channelDelegate != null) {
       webView.channelDelegate.onLoadStop(url);
@@ -370,7 +357,7 @@ public class InAppWebViewClient extends WebViewClient {
       credentialsProposed = CredentialDatabase.getInstance(view.getContext()).getHttpAuthCredentials(host, protocol, realm, port);
 
     URLCredential credentialProposed = null;
-    if (credentialsProposed != null && credentialsProposed.size() > 0) {
+    if (credentialsProposed != null && !credentialsProposed.isEmpty()) {
       credentialProposed = credentialsProposed.get(0);
     }
 
@@ -397,7 +384,7 @@ public class InAppWebViewClient extends WebViewClient {
               handler.proceed(username, password);
               break;
             case 2:
-              if (credentialsProposed.size() > 0) {
+              if (!credentialsProposed.isEmpty()) {
                 URLCredential credential = credentialsProposed.remove(0);
                 handler.proceed(credential.getUsername(), credential.getPassword());
               } else {
@@ -708,7 +695,7 @@ public class InAppWebViewClient extends WebViewClient {
     }
 
     WebResourceResponse response = null;
-    if (webView.contentBlockerHandler.getRuleList().size() > 0) {
+    if (!webView.contentBlockerHandler.getRuleList().isEmpty()) {
       try {
         response = webView.contentBlockerHandler.checkUrl(webView, request);
       } catch (Exception e) {